Fix brand module server-client split to resolve page hydration failure (auto via agent)
