+++ frontend/src/routes/+layout.server.ts
@@ -1,4 +1,4 @@
-import { loadBrandConfig } from '$lib/brand';
+import { loadBrandConfig } from '$lib/brand.server';
 import type { LayoutServerLoad } from './$types';
 
 export const load: LayoutServerLoad = async ({ locals }) => {
